moved bexioApi into partial classes
