+++ bexio.net/ApiBexio/Others/ApiBexio.Accounting.cs
@@ -1,0 +1,49 @@
+namespace bexio.net
+{
+	public partial class BexioApi
+	{
+		#region Accounting
+
+		// TODO
+
+		// Fetch a list of accounts
+		// Search Accounts
+
+		// Fetch a list of account groups
+
+		// Fetch a list of calendar years
+		// Create calendar year.
+		// Search calendar years
+		// Fetch a calendar year
+
+		// Fetch a list of business years
+		// Fetch a business year
+
+		// Fetch a list of currencies
+		// Create a currency
+		// Fetch a currency
+		// Delete a currency
+		// Update a currency
+		// Fetch exchange rates for currencies
+		// Fetch all possible currency codes
+
+		// Fetch a list of manual entries
+		// Create manual entry
+		// Get next reference number
+		// Fetch files of accounting entry line
+		// Add file to accounting entry line
+		// Fetch file of accounting entry line
+		// Delete file of accounting entry line
+
+		// Journal
+
+		// Fetch a list of taxes
+		// Fetch a tax
+		// Delete a tax
+
+		// Fetch a list of vat periods
+		// Fetch a vat period
+
+		#endregion
+	}
+}
