+++ bexio.net/ApiBexio/Others/ApiBexio.Banking.cs
@@ -1,0 +1,33 @@
+namespace bexio.net
+{
+	public partial class BexioApi
+	{
+		#region Banking
+		// TODO
+
+		// Fetch a list of bank accounts
+		// Fetch a single bank account
+
+		// Create IBAN payment
+		// Get IBAN payment
+		// Update IBAN payment
+
+		// Create ISR payment
+		// Get ISR payment
+		// Update ISR payment
+
+		// Create IS payment
+		// Get IS payment
+		// Update IS payment
+
+		// Create QR payment
+		// Get QR payment
+		// Update QR payment
+
+		// Fetch a list of payments
+		// Cancel a payment
+		// Delete a payment
+
+		#endregion
+	}
+}
