+++ bexio.net/ApiBexio/Others/ApiBexio.File.cs
@@ -1,0 +1,21 @@
+namespace bexio.net
+{
+	public partial class BexioApi
+	{
+		#region Files
+
+		// TODO
+
+		// Fetch a list of files
+		// Create new file
+		// Search files
+		// Get single file
+		// Delete a existing file
+		// Update existing file
+		// Download file
+		// Get file preview
+		// Show file usage
+
+		#endregion
+	}
+}
