+++ bexio.net/ApiBexio/Others/ApiBexio.Purchase.cs
@@ -1,0 +1,44 @@
+namespace bexio.net
+{
+	public partial class BexioApi
+	{
+		#region Purchase
+
+		// TODO
+
+		// Get Bills
+		// Create new Bill
+		// Get Bill
+		// Update Bill
+		// Delete Bill
+		// Update Bill status
+		// Execute Bill action
+		// Validate whether document number is available or not
+
+		// Get Expenses
+		// Create new Expense
+		// Get Expense
+		// Update Expense
+		// Delete Expense
+		// Update Expense status
+		// Execute Expense action
+		// Validate whether document number is available or not
+
+		// Fetch a list of purchase orders
+		// Create a purchase order
+		// Fetch a single purchase order
+		// Update a single purchase order
+		// Delete a purchase order
+
+		// Retrieve Outgoing Payments
+		// Create new Outgoing Payment
+		// Get Outgoing Payment
+		// Delete Outgoing Payment
+		// Retrieve Outgoing Payments
+		// Create new Outgoing Payment
+		// Get Outgoing Payment
+		// Delete Outgoing Payment
+
+		#endregion
+	}
+}
